% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExeFromPage}
\alias{getExeFromPage}
\title{get exe file}
\usage{
getExeFromPage(pageurl, dir, overwrite = FALSE)
}
\arguments{
\item{pageurl}{url of page}

\item{dir}{where to put file}

\item{overwrite}{overwrite existing file}
}
\value{
nothing
}
\description{
get a dispensing archive file
}
\details{
get one exe file
}
\author{
Barry Rowlingson
}

