% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLatestList}
\alias{getLatestList}
\title{Get latest list}
\usage{

  getLatestList(url = "http://www.hscic.gov.uk/searchcatalogue?q=title\%3a\%22presentation+level+data\%22&sort=Most+recent&size=100&page=1")
}
\arguments{
\item{url}{URL of listing}
}
\value{
vector of links to pages
}
\description{
get list of latest files
}
\details{
scrape web page for list
}
\author{
Barry Rowlingson
}

